add version info refactor deploy script
Deploy each version into different path. Before there was only the
latest version.

If on the exact tag version is git tag, functions are deployed to /functions/$tag path.
If thare are commits after tag tag is apended with commit hash, functions are deployed to /dev/$tag.
If repo is dirty username is added to the tag.

Deploy script flags:
  --only-cli                    just builds cli
  --use-old-functions-path      deploys functions to the /functions