refactor: remove util package
"My recommendation to improve the name of utils or helpers packages is
to analyse where they are imported and move the relevant functions into
the calling package. Even if this results in some code duplication this
is preferable to introducing an import dependency between two packages.
In the case where utility functions are used in many places, prefer
multiple packages, each focused on a single aspect with a
correspondingly descriptive name."

from: https://dave.cheney.net/2019/01/08/avoid-package-names-like-base-util-or-common