add vendedlogs prefix to api gateway logs
- added prefix mitigates errors such as:
BadRequestException: Cannot enable logging. Policy document length
breaking Cloudwatch Logs Constraints, either < 1 or > 5120

Explanation:

Api gateway must list each log group that they're sending logs to in the resource policy, and CloudWatch Logs resource policies are limited to 5120 characters. A service that sends logs to a large number of log groups may run into this limit.

To mitigate this, CloudWatch Logs monitors the size of resource policies used by the service that is sending logs, and when it detects that a policy approaches the size limit of 5120 characters, CloudWatch Logs automatically enables /aws/vendedlogs/* in the resource policy for that service. You can then start using log groups with names that start with /aws/vendedlogs/ as the destinations for logs from these services.